+++ gwexpy/types/__init__.py
@@ -1,29 +1,11 @@
 """gwexpy.types - Data type definitions and utilities."""
 
-from gwexpy.fields import ScalarField
-
 from .array import Array
 from .array2d import Array2D
 from .array3d import Array3D
 from .array4d import Array4D
 from .axis import AxisDescriptor, coerce_1d_quantity
 from .axis_api import AxisApiMixin
-from .field4d import Field4D
-from .field4d_collections import Field4DDict, Field4DList
-from .field4d_demo import (
-    make_demo_field4d,
-    make_propagating_gaussian,
-    make_sinusoidal_wave,
-    make_standing_wave,
-)
-from .field4d_signal import (
-    coherence_map,
-    compute_freq_space,
-    compute_psd,
-    compute_xcorr,
-    freq_space_map,
-    time_delay_map,
-)
 from .metadata import MetaData, MetaDataDict, MetaDataMatrix
 from .plane2d import Plane2D
 from .series_creator import as_series
@@ -48,23 +30,7 @@ __all__ = [
     "Plane2D",
     "Array3D",
     "Array4D",
-    "ScalarField",
-    "Field4D",
-    "Field4DList",
-    "Field4DDict",
     "TimePlaneTransform",
-    # Field4D demo
-    "make_demo_field4d",
-    "make_propagating_gaussian",
-    "make_sinusoidal_wave",
-    "make_standing_wave",
-    # Field4D signal processing
-    "compute_psd",
-    "freq_space_map",
-    "compute_freq_space",
-    "compute_xcorr",
-    "time_delay_map",
-    "coherence_map",
 ]
 
 # Dynamic import from gwpy (PEP 562)